feat(astro): make a new 'astro' module, including time, units, coordinates
