+++ caput/__init__.py
@@ -14,9 +14,7 @@ Submodules
     pipeline
     profile
     random
-    time
     tools
-    units
 """
 
 from importlib.metadata import PackageNotFoundError, version
