+++ caput/astro/__init__.py
@@ -1,0 +1,11 @@
+"""caput.astro.
+
+Astronomical utilities, including relevant coordinate transforms,
+time conversions, and units.
+"""
+
+from . import (
+    coordinates as coordinates,
+    time as time,
+    units as units,
+)
