+++ caput/astro/coordinates/__init__.py
@@ -1,0 +1,10 @@
+"""caput.coordinates.
+
+(Mostly) spherical coordinate transforms.
+"""
+
+from . import (
+    coord as coord,
+    spherical as spherical,
+    sphfunc as sphfunc,
+)
