--- caput/coordinates/__init__.py
@@ -1,10 +1,0 @@
-"""caput.coordinates.
-
-(Mostly) spherical coordinate transforms.
-"""
-
-from . import (
-    coord as coord,
-    spherical as spherical,
-    sphfunc as sphfunc,
-)
